+++ apps/examples/dpp/table.c
@@ -197,7 +197,7 @@ void table_init(int nsessions, struct am_event_alloc* alloc) {
     me->alloc = alloc;
 
     am_hsm_init(&me->hsm, am_hsm_state_make(table_initial));
-    am_ao_init(&me->ao, (am_ao_fn)am_hsm_start, (am_ao_fn)am_hsm_dispatch, me);
+    am_ao_init(&me->ao, am_hsm_start_cb, am_hsm_dispatch_cb, &me->hsm);
 }
 
 struct am_ao* table_get_obj(void) { return &m_table.ao; }